pcap-filter(7): Add note about ethernet-like hostnames. [skip ci]
Ethernet-like hostnames, i.e., hostnames that consist of exactly 12 hexadecimal characters, are parsed as Ethernet MAC addresses. If you try to match a hostname with this pattern, the error `ethernet address used in non-ether expression` is shown.
Unfortunately, those hostnames are quite common because Docker containers are by default named according to this template. Fixing the scanner to parse these hostnames correctly would destroy the handy option to drop `:` (or `.`, or `-`) in MAC addresses. Adding the note to the man page should at least help to find a work around that does not involve looking into the code.